Basic Styling on BookVersion modal + reorganize directory
- Reorganized directory to store all the Book version components in one folder.
- Added CSS Styling for readability of the modal.
