+++ src/pages/Study/BibleBookVersions/BookVersion.css
@@ -9,6 +9,7 @@
 
 .Books, .Versions {
     display: flex;
+    position: relative;
     padding: 5px 10px;
     align-items: center;
     justify-content: space-between;
