+++ src/pages/Study/BibleBookVersions/BookVersion.js
@@ -1,17 +1,20 @@
 import './BookVersion.css';
 import { GoTriangleDown } from "react-icons/go";
 import { useState } from 'react';
+import BookVersionComponent from './BookVersionComponent';
 
 const BookVersion = () => {
 
     const [book, setBook] = useState("Genesis");
     const [chapter, setChapter] = useState("Intro");
+    
 
     return (
         <div className="BookVersionHolder">
             <div className="Books">
                 <p className="BookNameDisplay">Genesis</p>
                 <GoTriangleDown className="BookVersionDownArrow"/>
+                <BookVersionComponent/>
             </div>
             <div className="Versions">
                 <p className="VersionNameDisplay">NIV</p>
